# Remove pg_stat_user_tables call and add toast relation metrics