--- a/postgres/changelog.d/17897.added
+++ b/postgres/changelog.d/17897.added
@@ -1,0 +1,1 @@
+Add toast metrics to relation metrics and remove usage of pg_stat_user_tables