### What does this PR do?
Replace our usage of pg_stat_user_tables by direct pg_stat_* function calls and add toast relation metrics.
It also applied the same check as the size query to exclude relations that are locked with an accessExclusiveLock.

### Motivation
pg_stat_user_tables view is inefficient on databases with a lot of tables, possibly leading to temporary bytes being read and written. To avoid this, we can directly use the pg_stat_* functions.

We also add the following metrics for the relation's toast tables:
- `postgresql.toast.index_scans`
- `postgresql.toast.rows_fetched`
- `postgresql.toast.rows_inserted`
- `postgresql.toast.rows_deleted`
- `postgresql.toast.live_rows`
- `postgresql.toast.dead_rows`
- `postgresql.toast.vacuumed`
- `postgresql.toast.autovacuumed`

We only generate metrics that are relevant:
- toast tables are only accessed through index so no need for seq scan metrics
- They are never updated, only deleted and inserted so no update and hot_update needed
- They are never analyzed, only vacuum and autovacuum is useful.

### Additional Notes

On a database with 98K tables and multiple indexes per tables:
```
EXPLAIN (ANALYZE, BUFFERS)
SELECT relname, schemaname,
       seq_scan, seq_tup_read,
       idx_scan, idx_tup_fetch,
       n_tup_ins, n_tup_upd, n_tup_del, n_tup_hot_upd, n_live_tup, n_dead_tup,
       vacuum_count, autovacuum_count,
       analyze_count, autoanalyze_count,
       EXTRACT(EPOCH FROM age(CURRENT_TIMESTAMP, last_vacuum)),
       EXTRACT(EPOCH FROM age(CURRENT_TIMESTAMP, last_autovacuum)),
       EXTRACT(EPOCH FROM age(CURRENT_TIMESTAMP, last_analyze)),
       EXTRACT(EPOCH FROM age(CURRENT_TIMESTAMP, last_autoanalyze))
FROM pg_stat_user_tables;
                                                                           QUERY PLAN
----------------------------------------------------------------------------------------------------------------------------------------------------------------
 Subquery Scan on pg_stat_all_tables  (cost=52857.22..98715.64 rows=193932 width=368) (actual time=291.975..1158.218 rows=98304 loops=1)
   Buffers: shared hit=32940, temp read=7024 written=7033
   I/O Timings: temp read=13.282 write=68.858
   ->  GroupAggregate  (cost=52857.22..90958.36 rows=193932 width=292) (actual time=291.971..1133.802 rows=98304 loops=1)
         Group Key: c.oid, n.nspname
         Buffers: shared hit=32940, temp read=7024 written=7033
         I/O Timings: temp read=13.282 write=68.858
         ->  Gather Merge  (cost=52857.22..75443.80 rows=193932 width=136) (actual time=291.880..417.681 rows=393216 loops=1)
               Workers Planned: 2
               Workers Launched: 2
               Buffers: shared hit=32940, temp read=7024 written=7033
               I/O Timings: temp read=13.282 write=68.858
               ->  Sort  (cost=51857.20..52059.21 rows=80805 width=136) (actual time=287.986..312.484 rows=131072 loops=3)
                     Sort Key: c.oid, n.nspname
                     Sort Method: external merge  Disk: 19176kB
                     Buffers: shared hit=32940, temp read=7024 written=7033
                     I/O Timings: temp read=13.282 write=68.858
                     Worker 0:  Sort Method: external merge  Disk: 18616kB
                     Worker 1:  Sort Method: external merge  Disk: 18400kB
                     ->  Parallel Hash Left Join  (cost=20183.72..42112.72 rows=80805 width=136) (actual time=101.616..208.383 rows=131072 loops=3)
                           Hash Cond: (c.oid = i.indrelid)
                           Buffers: shared hit=32784
                           ->  Hash Join  (cost=1536.20..22820.74 rows=80805 width=132) (actual time=32.645..102.723 rows=32768 loops=3)
                                 Hash Cond: (c.relnamespace = n.oid)
                                 Buffers: shared hit=18674
                                 ->  Parallel Seq Scan on pg_class c  (cost=0.00..21072.38 rows=80818 width=72) (actual time=0.007..55.327 rows=65572 loops=3)
                                       Filter: (relkind = ANY ('{r,t,m,p}'::"char"[]))
                                       Rows Removed by Filter: 163941
                                       Buffers: shared hit=16769
                                 ->  Hash  (cost=1126.60..1126.60 rows=32768 width=68) (actual time=32.518..32.519 rows=32771 loops=3)
                                       Buckets: 65536 (originally 32768)  Batches: 1 (originally 1)  Memory Usage: 3713kB
                                       Buffers: shared hit=1905
                                       ->  Seq Scan on pg_namespace n  (cost=0.00..1126.60 rows=32768 width=68) (actual time=0.056..24.289 rows=32771 loops=3)
                                             Filter: ((nspname <> ALL ('{pg_catalog,information_schema}'::name[])) AND (nspname !~ '^pg_toast'::text))
                                             Rows Removed by Filter: 4
                                             Buffers: shared hit=1905
                           ->  Parallel Hash  (cost=16086.67..16086.67 rows=204868 width=8) (actual time=68.208..68.209 rows=163894 loops=3)
                                 Buckets: 524288  Batches: 1  Memory Usage: 23424kB
                                 Buffers: shared hit=14038
                                 ->  Parallel Seq Scan on pg_index i  (cost=0.00..16086.67 rows=204868 width=8) (actual time=0.007..30.057 rows=163894 loops=3)
                                       Buffers: shared hit=14038
 Planning:
   Buffers: shared hit=497
 Planning Time: 1.451 ms
 Execution Time: 1169.630 ms
```

With the new plan: 
```
EXPLAIN (ANALYZE, BUFFERS)
SELECT
  N.nspname, C.relname,
  pg_stat_get_numscans(C.oid), pg_stat_get_tuples_returned(C.oid),
  I.idx_scan, I.idx_tup_fetch,
  pg_stat_get_tuples_inserted(C.oid), pg_stat_get_tuples_updated(C.oid), pg_stat_get_tuples_deleted(C.oid), pg_stat_get_tuples_hot_updated(C.oid),
  pg_stat_get_live_tuples(C.oid), pg_stat_get_dead_tuples(C.oid),
  pg_stat_get_vacuum_count(C.oid), pg_stat_get_autovacuum_count(C.oid),
  pg_stat_get_analyze_count(C.oid), pg_stat_get_autoanalyze_count(C.oid),
  EXTRACT(EPOCH FROM age(CURRENT_TIMESTAMP, pg_stat_get_last_vacuum_time(C.oid))),
  EXTRACT(EPOCH FROM age(CURRENT_TIMESTAMP, pg_stat_get_last_autovacuum_time(C.oid))),
  EXTRACT(EPOCH FROM age(CURRENT_TIMESTAMP, pg_stat_get_last_analyze_time(C.oid))),
  EXTRACT(EPOCH FROM age(CURRENT_TIMESTAMP, pg_stat_get_last_autoanalyze_time(C.oid))),
  pg_stat_get_numscans(idx_toast.indexrelid),
  pg_stat_get_tuples_fetched(idx_toast.indexrelid),
  pg_stat_get_tuples_inserted(C.reltoastrelid),
  pg_stat_get_tuples_deleted(C.reltoastrelid),
  pg_stat_get_live_tuples(C.reltoastrelid),
  pg_stat_get_dead_tuples(C.reltoastrelid),
  pg_stat_get_vacuum_count(C.reltoastrelid),
  pg_stat_get_autovacuum_count(C.reltoastrelid)
FROM pg_class C
LEFT JOIN pg_namespace N ON (N.oid = C.relnamespace)
LEFT JOIN pg_index idx_toast ON (idx_toast.indrelid = C.reltoastrelid)
LEFT JOIN pg_locks L ON C.oid = L.relation AND L.locktype = 'relation'
LEFT JOIN LATERAL (
    SELECT sum(pg_stat_get_numscans(indexrelid))::bigint AS idx_scan,
           sum(pg_stat_get_tuples_fetched(indexrelid))::bigint AS idx_tup_fetch
      FROM pg_index
     WHERE pg_index.indrelid = C.oid) I ON true
WHERE C.relkind = 'r'
    AND NOT (nspname = ANY('{{pg_catalog,information_schema}}'))
    AND (L.relation IS NULL OR L.mode <> 'AccessExclusiveLock' OR NOT L.granted);

                                                                            QUERY PLAN
------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Nested Loop Left Join  (cost=1472.84..624124.10 rows=373898 width=432) (actual time=15.007..1934.766 rows=98304 loops=1)
   Buffers: shared hit=911806
   ->  Nested Loop Left Join  (cost=1467.30..36758.88 rows=373898 width=140) (actual time=14.970..584.497 rows=98304 loops=1)
         Buffers: shared hit=410620
         ->  Hash Join  (cost=1466.87..27465.91 rows=97615 width=136) (actual time=14.950..249.973 rows=98304 loops=1)
               Hash Cond: (c.relnamespace = n.oid)
               Buffers: shared hit=17404
               ->  Hash Left Join  (cost=12.56..25755.32 rows=97621 width=76) (actual time=1.455..195.083 rows=98372 loops=1)
                     Hash Cond: (c.oid = l.relation)
                     Filter: ((l.relation IS NULL) OR (l.mode <> 'AccessExclusiveLock'::text) OR (NOT l.granted))
                     Buffers: shared hit=16769
                     ->  Seq Scan on pg_class c  (cost=0.00..25375.75 rows=97864 width=76) (actual time=0.007..168.528 rows=98372 loops=1)
                           Filter: (relkind = 'r'::"char")
                           Rows Removed by Filter: 590168
                           Buffers: shared hit=16769
                     ->  Hash  (cost=12.50..12.50 rows=5 width=37) (actual time=1.442..1.443 rows=11 loops=1)
                           Buckets: 1024  Batches: 1  Memory Usage: 9kB
                           ->  Function Scan on pg_lock_status l  (cost=0.00..12.50 rows=5 width=37) (actual time=1.434..1.437 rows=11 loops=1)
                                 Filter: (locktype = 'relation'::text)
                                 Rows Removed by Filter: 2
               ->  Hash  (cost=1044.66..1044.66 rows=32771 width=68) (actual time=13.471..13.472 rows=32773 loops=1)
                     Buckets: 65536  Batches: 1  Memory Usage: 3713kB
                     Buffers: shared hit=635
                     ->  Seq Scan on pg_namespace n  (cost=0.00..1044.66 rows=32771 width=68) (actual time=0.009..7.349 rows=32773 loops=1)
                           Filter: (nspname <> ALL ('{{pg_catalog,information_schema}}'::name[]))
                           Rows Removed by Filter: 2
                           Buffers: shared hit=635
         ->  Memoize  (cost=0.43..0.66 rows=4 width=8) (actual time=0.003..0.003 rows=1 loops=98304)
               Cache Key: c.reltoastrelid
               Cache Mode: logical
               Hits: 0  Misses: 98304  Evictions: 0  Overflows: 0  Memory Usage: 10368kB
               Buffers: shared hit=393216
               ->  Index Scan using pg_index_indrelid_index on pg_index idx_toast  (cost=0.42..0.65 rows=4 width=8) (actual time=0.002..0.002 rows=1 loops=98304)
                     Index Cond: (indrelid = c.reltoastrelid)
                     Buffers: shared hit=393216
   ->  Memoize  (cost=5.55..5.57 rows=1 width=16) (actual time=0.009..0.009 rows=1 loops=98304)
         Cache Key: c.oid
         Cache Mode: binary
         Hits: 0  Misses: 98304  Evictions: 0  Overflows: 0  Memory Usage: 11136kB
         Buffers: shared hit=501186
         ->  Aggregate  (cost=5.54..5.55 rows=1 width=16) (actual time=0.008..0.008 rows=1 loops=98304)
               Buffers: shared hit=501186
               ->  Index Scan using pg_index_indrelid_index on pg_index  (cost=0.42..5.49 rows=4 width=4) (actual time=0.002..0.003 rows=4 loops=98304)
                     Index Cond: (indrelid = c.oid)
                     Buffers: shared hit=501186
 Planning:
   Buffers: shared hit=17
 Planning Time: 0.440 ms
 Execution Time: 1943.374 ms
```

While the query execution is slower:
- It's not doing a parallel query (less overhead on the db)
- There's no I/O generated (no disk overhead)
- It needs to do additional work to generate the toast values (join on the pg_index for the toast idx metrics)

Doing this test on a db with less tables (30K), we have:

Old query:
```
                                                                         QUERY PLAN
-------------------------------------------------------------------------------------------------------------------------------------------------------------
 Subquery Scan on pg_stat_all_tables  (cost=7161.67..13722.20 rows=27744 width=368) (actual time=30.458..81.321 rows=30004 loops=1)
   Buffers: shared hit=5528
   ->  GroupAggregate  (cost=7161.67..12612.44 rows=27744 width=316) (actual time=30.446..75.657 rows=30004 loops=1)
         Group Key: c.oid, n.nspname
         Buffers: shared hit=5528
         ->  Gather Merge  (cost=7161.67..10392.92 rows=27744 width=136) (actual time=30.424..37.729 rows=30004 loops=1)
               Workers Planned: 2
               Workers Launched: 2
               Buffers: shared hit=5528
               ->  Sort  (cost=6161.65..6190.55 rows=11560 width=136) (actual time=27.851..28.823 rows=10001 loops=3)
                     Sort Key: c.oid, n.nspname
                     Sort Method: quicksort  Memory: 2135kB
                     Buffers: shared hit=5528
                     Worker 0:  Sort Method: quicksort  Memory: 1065kB
                     Worker 1:  Sort Method: quicksort  Memory: 2215kB
                     ->  Hash Left Join  (cost=1360.41..5381.53 rows=11560 width=136) (actual time=10.554..24.628 rows=10001 loops=3)
                           Hash Cond: (c.oid = i.indrelid)
                           Buffers: shared hit=5386
                           ->  Hash Join  (cost=1.08..3879.11 rows=11560 width=132) (actual time=0.127..11.517 rows=10001 loops=3)
                                 Hash Cond: (c.relnamespace = n.oid)
                                 Buffers: shared hit=3346
                                 ->  Parallel Seq Scan on pg_class c  (cost=0.00..3753.03 rows=23120 width=72) (actual time=0.007..8.742 rows=20038 loops=3)
                                       Filter: (relkind = ANY ('{r,t,m,p}'::"char"[]))
                                       Rows Removed by Filter: 10104
                                       Buffers: shared hit=3233
                                 ->  Hash  (cost=1.06..1.06 rows=2 width=68) (actual time=0.047..0.048 rows=1 loops=3)
                                       Buckets: 1024  Batches: 1  Memory Usage: 9kB
                                       Buffers: shared hit=3
                                       ->  Seq Scan on pg_namespace n  (cost=0.00..1.06 rows=2 width=68) (actual time=0.041..0.042 rows=1 loops=3)
                                             Filter: ((nspname <> ALL ('{pg_catalog,information_schema}'::name[])) AND (nspname !~ '^pg_toast'::text))
                                             Rows Removed by Filter: 3
                                             Buffers: shared hit=3
                           ->  Hash  (cost=981.92..981.92 rows=30192 width=8) (actual time=10.336..10.337 rows=30168 loops=3)
                                 Buckets: 32768  Batches: 1  Memory Usage: 1435kB
                                 Buffers: shared hit=2040
                                 ->  Seq Scan on pg_index i  (cost=0.00..981.92 rows=30192 width=8) (actual time=0.014..5.518 rows=30168 loops=3)
                                       Buffers: shared hit=2040
 Planning:
   Buffers: shared hit=21
 Planning Time: 0.502 ms
 Execution Time: 82.535 ms
```

New query:
```
                                                                    QUERY PLAN
--------------------------------------------------------------------------------------------------------------------------------------------------
 Nested Loop Left Join  (cost=1381.28..122830.86 rows=13838 width=432) (actual time=6.906..140.366 rows=30004 loops=1)
   Buffers: shared hit=63925
   ->  Hash Left Join  (cost=1372.96..6107.33 rows=13838 width=140) (actual time=6.881..50.827 rows=30004 loops=1)
         Hash Cond: (c.reltoastrelid = idx_toast.indrelid)
         Buffers: shared hit=3914
         ->  Hash Left Join  (cost=13.64..4488.61 rows=13830 width=136) (actual time=0.233..35.418 rows=30004 loops=1)
               Hash Cond: (c.oid = l.relation)
               Filter: ((l.relation IS NULL) OR (l.mode <> 'AccessExclusiveLock'::text) OR (NOT l.granted))
               Buffers: shared hit=3234
               ->  Hash Join  (cost=1.07..4424.04 rows=13864 width=136) (actual time=0.035..30.852 rows=30004 loops=1)
                     Hash Cond: (c.relnamespace = n.oid)
                     Buffers: shared hit=3234
                     ->  Seq Scan on pg_class c  (cost=0.00..4273.06 rows=27729 width=76) (actual time=0.013..25.509 rows=30072 loops=1)
                           Filter: (relkind = 'r'::"char")
                           Rows Removed by Filter: 60352
                           Buffers: shared hit=3233
                     ->  Hash  (cost=1.05..1.05 rows=2 width=68) (actual time=0.009..0.010 rows=2 loops=1)
                           Buckets: 1024  Batches: 1  Memory Usage: 9kB
                           Buffers: shared hit=1
                           ->  Seq Scan on pg_namespace n  (cost=0.00..1.05 rows=2 width=68) (actual time=0.005..0.006 rows=2 loops=1)
                                 Filter: (nspname <> ALL ('{{pg_catalog,information_schema}}'::name[]))
                                 Rows Removed by Filter: 2
                                 Buffers: shared hit=1
               ->  Hash  (cost=12.50..12.50 rows=5 width=37) (actual time=0.195..0.195 rows=11 loops=1)
                     Buckets: 1024  Batches: 1  Memory Usage: 9kB
                     ->  Function Scan on pg_lock_status l  (cost=0.00..12.50 rows=5 width=37) (actual time=0.189..0.190 rows=11 loops=1)
                           Filter: (locktype = 'relation'::text)
                           Rows Removed by Filter: 1
         ->  Hash  (cost=981.92..981.92 rows=30192 width=8) (actual time=6.624..6.624 rows=30168 loops=1)
               Buckets: 32768  Batches: 1  Memory Usage: 1435kB
               Buffers: shared hit=680
               ->  Seq Scan on pg_index idx_toast  (cost=0.00..981.92 rows=30192 width=8) (actual time=0.006..3.397 rows=30168 loops=1)
                     Buffers: shared hit=680
   ->  Aggregate  (cost=8.32..8.34 rows=1 width=16) (actual time=0.001..0.001 rows=1 loops=30004)
         Buffers: shared hit=60011
         ->  Index Scan using pg_index_indrelid_index on pg_index  (cost=0.29..8.30 rows=1 width=4) (actual time=0.001..0.001 rows=0 loops=30004)
               Index Cond: (indrelid = c.oid)
               Buffers: shared hit=60011
 Planning:
   Buffers: shared hit=11
 Planning Time: 0.429 ms
 Execution Time: 141.498 ms
```

Same observations, while slower, we avoid the parallel query.

### Review checklist (to be filled by reviewers)

- [ ] Feature or bugfix MUST have appropriate tests (unit, integration, e2e)
- [ ] [Changelog entries](https://datadoghq.dev/integrations-core/guidelines/pr/#changelog-entries) must be created for modifications to shipped code
- [ ] Add the `qa/skip-qa` label if the PR doesn't need to be tested during QA.
- [ ] If you need to backport this PR to another branch, you can add the `backport/<branch-name>` label to the PR and it will automatically open a backport PR once this one is merged