@bonnefoa 2024-06-27
Related task: [ADPPOS-982]